Add custom ui theme

+++ scripts/IntegrityManager.cs
@@ -30,7 +30,7 @@ public partial class IntegrityManager : Node
         ulong users = UserManager.Instance.Users;
         uint servers = UpgradeManager.Instance.CheckUpgrade(Upgrade.Servers);
         uint supportedUsers = (uint)Math.Pow(10, servers + 2);
-        float serverFactor = 1.0f - (users / supportedUsers);
+        float serverFactor = 1.0f - ((float)users / (float)supportedUsers);
         this.Integrity = ddosFactor * Math.Max(serverFactor, 0.0f);
     }
 }
